Created "gameSelect" function
Wrote function 3. Instructions were commented at the top of the code. Hardcoded a list of teams from the Big 12 (BYU's conference), but the function will take a list as a parameter if need be. Returns a home team name and an away team name from that list based on user input. Didn't code any exception handling in there but I can add that if you guys want.