Moved the name input to function
In the instructions it says to ask for the name in the function so I just move the line into the function and adjusted so that it just does all of that as part of the function and then assigns the name to a variable. We don't end up using the name again after that but hey, whatever the instructions say.